Added Graphics abstraction layer
This wraps around all the calls I was making to Horde3D. I did this
because I'm thinking it's almost certain I'll need to use something
other than Horde3D for graphics eventually.

I renamed/moved the Main file in preparation for merging this to
master. Unfortunately, this does not result in a very clean merge.